website_vacations/backend/requirements.txt
psycopg2-binary==2.9.10
pytest==8.3.3
python-dotenv==1.0.1
typing-extensions>=4.9.0
flask==3.0.3
flask-cors==5.0.0


